Moved some cost matrix cmds into their own files
